// frontend/src/app/core/api/model/exam-timetable-response.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ExamTimetableResponse { 
    id?: number;
    name?: string;
    classGroupId?: number;
    classGroupName?: string;
    periodId?: number;
    periodLabel?: string;
    examSessionId?: number;
    examSessionLabel?: string;
    sessionType?: ExamTimetableResponse.SessionTypeEnum;
    status?: ExamTimetableResponse.StatusEnum;
    createdDate?: string;
    modifiedDate?: string;
}
export namespace ExamTimetableResponse {
    export const SessionTypeEnum = {
        Main: 'MAIN',
        Retake: 'RETAKE'
    } as const;
    export type SessionTypeEnum = typeof SessionTypeEnum[keyof typeof SessionTypeEnum];
    export const StatusEnum = {
        Draft: 'DRAFT',
        Published: 'PUBLISHED'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}